+++ Assets/Prefabs/Node.prefab
@@ -263,7 +263,7 @@ GameObject:
   serializedVersion: 5
   m_Component:
   - component: {fileID: 4732136330890278}
-  - component: {fileID: 114484797548686436}
+  - component: {fileID: 114540608570995662}
   m_Layer: 8
   m_Name: Node
   m_TagString: Untagged
@@ -1046,7 +1046,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
@@ -1133,7 +1133,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
@@ -1176,21 +1176,6 @@ MonoBehaviour:
     layerIndex: 0
   _eventTable:
     _entries: []
---- !u!114 &114484797548686436
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1885264354301550}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b6a0ccd4bff468d46858de493ceda7bb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  parent: {fileID: 0}
-  children: []
-  text: Unity
-  cube: {fileID: 1942337001675556}
 --- !u!114 &114501002478848566
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1264,7 +1249,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
@@ -1278,6 +1263,28 @@ MonoBehaviour:
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
   axis: 1
+--- !u!114 &114540608570995662
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1885264354301550}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3cde2cc3ef0c4a344ba620cd94488b56, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  data:
+    id: 0
+    text: Unity
+    xPos: 0
+    yPos: 0
+    zPos: 0
+    parentId: -1
+    childrenIds: 
+  parent: {fileID: 0}
+  children: []
+  cube: {fileID: 1942337001675556}
 --- !u!114 &114557475436284342
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1363,7 +1370,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
@@ -1529,7 +1536,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
@@ -1616,7 +1623,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 114484797548686436}
+      - m_Target: {fileID: 114540608570995662}
         m_MethodName: rePostion
         m_Mode: 1
         m_Arguments:
