+++ Assets/Scripts/Serialization.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 52bbaa73e635fab43a6d994227a09c1c
+folderAsset: yes
+timeCreated: 1545052226
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
