+++ Assets/Scripts/Serialization/NodeData.cs
@@ -4,18 +4,18 @@ using System;
 using System.Collections.Generic;
 
 [Serializable]
-public class NodeData {
+public class NodeData{
     /*
      The NodeData class is a serializable object that
      can hold data in a specified and serializable form.
      This way, data that defines the Min3dmap can be
      saved to a file and be converted back into a Min3dmap.
      */
-    public int id;
-    public string text;
+    public int id = 0;
+    public string text = "Unity";
     public float xPos;
     public float yPos;
     public float zPos;
-    public int parentId;
-    public List<int> childrenIds;
+    public int parentId = -1;
+    //public List<int> childrenIds = new List<int>();
 }
